-A big improvement by removed the flushAuthenticatedData  function from  the InterfaceAuthDataSource. The implementation of the framework is more  error prone and easier.
Redesigned the PHP Globally reachabel values handling subsystem
Inherited the  session handling subsystem form the last mentioned.

This made the whole workflow to get completely unit tested,
The but it does not mean that the functional testing should not be
alligned with the refactoring.

Removed some bad behaviors of the captcha, there was some
misconception how it released into the process.

More documentation, yes, and a little proud of it.